Finilaze unit tests.

+++ internal/models/transaction.go
@@ -10,20 +10,20 @@ import (
 type Transaction struct {
 	InvoiceID   int     `json:"invoice_id" binding:"required" example:"12"`
 	Amount      int32   `json:"-" binding:"required" example:"49297"`
-	AmountFloat float32 `json:"amount" example:"956.32"`
+	AmountFloat float64 `json:"amount" example:"956.32"`
 	Reference   string  `json:"reference" binding:"required" example:"JMPINV200220117"`
 }
 
 // TransactionJSON map the info about an transaction
 type TransactionJSON struct {
 	InvoiceID   int     `json:"invoice_id" binding:"required" example:"12"`
-	AmountFloat float32 `json:"amount" binding:"required" example:"956.32"`
+	AmountFloat float64 `json:"amount" binding:"required" example:"956.32"`
 	Reference   string  `json:"reference" binding:"required" example:"JMPINV200220117"`
 }
 
 // AfterFind overload balance to fit to specs
 func (i *Transaction) AfterFind(tx *gorm.DB) (err error) {
-	i.AmountFloat = float32(i.Amount) / 100.0
+	i.AmountFloat = float64(i.Amount) / 100.0
 	return
 }
